Add Person and Shift entities, and implement repositories for database interaction

+++ src/main/java/com/scheduler/schedulerapp/repository/ShiftRepository.java
@@ -3,5 +3,5 @@ package com.scheduler.schedulerapp.repository;
 import com.scheduler.schedulerapp.model.Shift;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShiftRepository extends JpaRepository<Shift, Integer> {
+public interface ShiftRepository extends JpaRepository<Shift, Long> {
 }
